% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_timeseries.plot.R
\name{ceg_timeseries_plot}
\alias{ceg_timeseries_plot}
\title{ceg_timeseries_plot}
\usage{
ceg_timeseries_plot(
  data,
  variable = "TChla",
  station = "P3",
  res_x = 500,
  res_y = 500,
  colorbar = "auto",
  palette = colorRampPalette(c("#00008F", "#0000FF", "#004FFF", "#00FFFF", "#80FF80",
    "#FFFF00", "#FFAF00", "#FF0000", "#800000")),
  mincolor,
  maxcolor,
  shape = 16,
  point_size = 2,
  text_size = 4,
  base_size = 16,
  family = "sans",
  save_width = 12,
  save_height = 6,
  format = "pdf"
)
}
\arguments{
\item{data}{A data.frame. including Lon, Lat, Station, Time, Variable, Value, Unit}

\item{variable}{character. unique(data$Variable). Default is "P3"}

\item{station}{character. unique(data$Station). Default is "Section1"}

\item{res_x}{A number. Interpolation resolution, Default is 500}

\item{res_y}{A number. Interpolation resolution, Default is 500}

\item{colorbar}{character."auto" or "manual"}

\item{palette}{A color vector for data$Value. Default is colorRampPalette(c("#00008F","#0000FF","#004FFF", "#00FFFF", "#80FF80", "#FFFF00", "#FFAF00","#FF0000","#800000"))}

\item{mincolor}{A number. If colorbar = "manual", need a mincolor number}

\item{maxcolor}{A number. If colorbar = "manual", need a maxcolor number}

\item{shape}{A number. Shape for Sample points, Default is 16}

\item{point_size}{A number. Size for Sample points, Default is 2}

\item{text_size}{A number. Size for text contour, Default is 4}

\item{base_size}{A number. Base size, default is 16}

\item{family}{character. "sans"/ Arial, "serif"/Time new roman and any other fronts, Default is "sans"}

\item{save_width}{A number. width for save figures, Default is 10}

\item{save_height}{A number. height for save figures, Default is 8}

\item{format}{character. file type for save figures, Default is "pdf"}
}
\value{
A figure
}
\description{
ceg_timeseries_plot
}
\examples{
NULL
}
\author{
Weinan Li
}
