% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map_colors.R
\name{get_map_colors}
\alias{get_map_colors}
\title{get_map_colors}
\usage{
get_map_colors(
  lon_min = 100,
  lon_max = 125,
  lat_min = 0,
  lat_max = 25,
  nbreaks = 100,
  oceancolor = rev(RColorBrewer::brewer.pal(9, "Blues")),
  landcolor = RColorBrewer::brewer.pal(11, "PiYG")
)
}
\arguments{
\item{lon_min}{A number. Default is 100}

\item{lon_max}{A number. Default is 125}

\item{lat_min}{A number. Default is 0}

\item{lat_max}{A number. Default is 25}

\item{nbreaks}{one number. Default is 100}

\item{oceancolor}{A color vector for ocean. Default is rev(RColorBrewer::brewer.pal(9, "Blues"))}

\item{landcolor}{A color vector for land. Default is RColorBrewer::brewer.pal(11, "PiYG")}
}
\value{
A data.table data.frame
}
\description{
get_map_colors
}
\examples{
mapcolors <- get_map_colors()
}
\author{
Weinan Li
}
